Add Gaussian Blur
Currently using a borrowed class for blurring
PImages in Processing. The class is ugly and
it mutates the image directly. As such, I have
wrapped this implementation so that it is
used by the new Blur Object with Blur.gaussian
that returns a new image without modifying the source
image